+++ src/main/java/com/enonic/harvest/harvestclient/models/DayEntry.java
@@ -3,10 +3,6 @@ package com.enonic.harvest.harvestclient.models;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import java.math.BigDecimal;
 import java.util.Date;
 
@@ -19,22 +15,22 @@ public class DayEntry
     private BigDecimal hours;
 
     @JsonProperty(value = "id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty(value = "notes")
     private String notes;
 
     @JsonProperty(value = "project_id")
-    private Integer projectId;
+    private Long projectId;
 
     @JsonProperty(value = "spent_at")
     private Date spentAt;
 
     @JsonProperty(value = "task_id")
-    private Integer taskId;
+    private Long taskId;
 
     @JsonProperty(value = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @JsonProperty(value = "is_billed")
     private boolean isBilled;
@@ -68,12 +64,12 @@ public class DayEntry
         this.hours = hours;
     }
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
@@ -98,12 +94,12 @@ public class DayEntry
         this.notes = notes;
     }
 
-    public Integer getProjectId()
+    public Long getProjectId()
     {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId)
+    public void setProjectId(Long projectId)
     {
         this.projectId = projectId;
     }
@@ -118,12 +114,12 @@ public class DayEntry
         this.spentAt = spentAt;
     }
 
-    public Integer getTaskId()
+    public Long getTaskId()
     {
         return taskId;
     }
 
-    public void setTaskId(Integer taskId)
+    public void setTaskId(Long taskId)
     {
         this.taskId = taskId;
     }
@@ -138,12 +134,12 @@ public class DayEntry
         this.updatedAt = updatedAt;
     }
 
-    public Integer getUserId()
+    public Long getUserId()
     {
         return userId;
     }
 
-    public void setUserId(Integer userId)
+    public void setUserId(Long userId)
     {
         this.userId = userId;
     }
