+++ src/main/java/com/enonic/harvest/harvestclient/models/Project.java
@@ -6,10 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.Unmarshaller;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import java.io.InputStream;
 import java.math.BigDecimal;
 import java.util.Date;
@@ -23,10 +19,10 @@ public class Project
     private String name;
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("client_id")
-    private Integer clientId;
+    private Long clientId;
 
     @JsonProperty("code")
     private String code;
@@ -115,22 +111,22 @@ public class Project
         this.name = name;
     }
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
 
-    public Integer getClientId()
+    public Long getClientId()
     {
         return clientId;
     }
 
-    public void setClientId(Integer clientId)
+    public void setClientId(Long clientId)
     {
         this.clientId = clientId;
     }
