+++ src/main/java/com/enonic/harvest/harvestclient/models/Task.java
@@ -6,10 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.Unmarshaller;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import java.io.InputStream;
 import java.math.BigDecimal;
 import java.util.Date;
@@ -27,7 +23,7 @@ public class Task
     private BigDecimal defaultHourlyRate;
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("is_default")
     private boolean isDefault;
@@ -71,12 +67,12 @@ public class Task
         this.defaultHourlyRate = defaultHourlyRate;
     }
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
