+++ src/main/java/com/enonic/harvest/harvestclient/models/UserAssignment.java
@@ -3,9 +3,6 @@ package com.enonic.harvest.harvestclient.models;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
 import java.math.BigDecimal;
 import java.util.Date;
 
@@ -13,13 +10,13 @@ import java.util.Date;
 public class UserAssignment
 {
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("user_id")
-    private Integer userId;
+    private Long userId;
 
     @JsonProperty("project_id")
-    private Integer projectId;
+    private Long projectId;
 
     @JsonProperty("deactivated")
     private boolean deactivated;
@@ -36,32 +33,32 @@ public class UserAssignment
     @JsonProperty("created_at")
     private Date createdAt;
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
 
-    public Integer getUserId()
+    public Long getUserId()
     {
         return userId;
     }
 
-    public void setUserId(Integer userId)
+    public void setUserId(Long userId)
     {
         this.userId = userId;
     }
 
-    public Integer getProjectId()
+    public Long getProjectId()
     {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId)
+    public void setProjectId(Long projectId)
     {
         this.projectId = projectId;
     }
